requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
motor==3.6.0
pydantic-settings==2.6.1
email-validator==2.2.0

